EREF-64 add Java test that implements identifier-query-test-cases
Note that the schema identifier-query-test-cases.xsd refers to such
types that are already used for code generation in the commons module.
We want to use those commons module classes but there is no way to
tell code generation to not generate said classes. Let's just remove
the duplicate classes after code generation.