EREF-27 allow access to gate's platform REST api via pre-auth headers
Gate expects to get headers that identify the calling client and its
role. These headers must be added by some other component, that is
responsible for authenticating the client. This component may be,
for example, a proxy like nginx.

For local development, configure the Apache proxy to accept headers
X-Mock-Pre-Authenticated-User-Id and X-Mock-Pre-Authenticated-User-Role
that may be used to authenticate the client.