EREF-73 fix gate config to support running it directly from sources
Fix the config so that when running the gate from sources as a Java
app running on the host (and not in docker container), the gate:

 * can connect to the rabbitmq container
 * logs to stdout and not to files specific to docker container

Note that overriding logback configuration file via Spring Boot
properties did not work when default filename was used for the
default config. So, rename the default config to logback-default.xml
and the simplified version logback-dev.xml and choose appropriate
file in Spring Boot properties.